+++ tests/LuyaComposerPluginTest.php
@@ -2,7 +2,6 @@
 
 namespace luya\composer\tests;
 
-use Composer\TestCase;
 use luya\composer\Plugin;
 use Composer\Installer\PackageEvent;
 use Composer\Script\Event;
@@ -13,46 +12,43 @@ class LuyaComposerPluginTest extends TestCase
      * @var Plugin
      */
     protected $plugin;
-    /**
-     * @var Composer
-     */
-    protected $composer;
-    /**
-     * @var IOInterface
-     */
-    protected $io;
-    /**
-     * @var \PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $package;
     
     protected function setUp()
     {
+        parent::setUp();
+        
         $this->plugin = new Plugin();
-        $this->composer = $this->getMockBuilder('Composer\Composer')->getMock();
-        $this->io = $this->getMockBuilder('Composer\IO\IOInterface')->getMock();
-        $this->package = $this->getMockBuilder('Composer\Package\RootPackageInterface')->getMock();
+        $this->plugin->linkPath = __DIR__ . '/data/luya';
     }
     
     protected function tearDown()
     {
         $this->plugin = null;
-        $this->composer = null;
-        $this->io = null;
+        parent::tearDown();
     }
     
     public function testSubscribeEvents()
     {
         $this->plugin->activate($this->composer, $this->io);
         
-        $this->assertCount(3, $this->plugin->getSubscribedEvents());
-        
-        $packageEvent = $this->getMockBuilder(PackageEvent::class)->disableOriginalConstructor()->getMock();
-        
-        $this->plugin->findCoreRepo($packageEvent);
-        
-        $scriptEvent = $this->getMockBuilder(Event::class)->disableOriginalConstructor()->getMock();
-        
-        $this->plugin->postUpdateScript($scriptEvent);
+        $this->assertCount(4, $this->plugin->getSubscribedEvents());
     }
+    
+    public function testPostUpdateScript()
+    {
+        $this->invokeSetProperty($this->plugin, '_packageInstalls', [
+            'luyadev/luya-module-admin',
+            'luyadev/luya-core',
+            'luyadev/luya-foo',
+        ]);
+    
+        $scriptEvent = new Event('post-update', $this->composer, $this->io);
+    
+        $this->plugin->postUpdateScript($scriptEvent);
+    
+        $luyaLinkTarget = @readlink(__DIR__ . '/data/luya');
+        $this->assertNotFalse($luyaLinkTarget, 'Luya file link missing.');
+        $this->assertStringStartsNotWith('/', $luyaLinkTarget, 'Link target should not be a absolute path.');
+    }
+    
 }